Update README.md
I encountered an issue while using the Spotify Developer account. It seems that the API only provides the genre information for artists, not for specific songs. Unfortunately, this makes the genre data unusable for my needs, as I require genre information tied to individual tracks.

To address this, I decided to remove the parts of the code that relied on or connected to genre data. Let me know if you have suggestions or alternative solutions for this limitation.